refactor: unify Api services with ApiService interface
