--- app/src/main/java/com/omouravictor/ratesnow/data/network/hgbrasil/rates/CurrencyApi.kt
@@ -1,12 +1,0 @@
-package com.omouravictor.ratesnow.data.network.hgbrasil.rates
-
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface CurrencyApi {
-
-    @GET("finance")
-    suspend fun getCurrencies(
-        @Query("fields") field: String
-    ): SourceRequestCurrencyModel
-}
