--- app/src/main/java/com/omouravictor/ratesnow/data/network/hgbrasil/stock/StocksApi.kt
@@ -1,14 +1,0 @@
-package com.omouravictor.ratesnow.data.network.hgbrasil.stock
-
-import retrofit2.Response
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface StocksApi {
-
-    @GET("finance")
-    suspend fun getStocks(
-        @Query("fields") field: String,
-        @Query("key") apikey: String
-    ): Response<SourceRequestStockModel>
-}
