+++ app/src/main/java/com/omouravictor/ratesnow/data/repository/BitCoinRepository.kt
@@ -1,20 +1,19 @@
 package com.omouravictor.ratesnow.data.repository
 
-import com.omouravictor.ratesnow.BuildConfig
 import com.omouravictor.ratesnow.data.local.AppDataBase
 import com.omouravictor.ratesnow.data.local.entity.BitCoinEntity
-import com.omouravictor.ratesnow.data.network.hgbrasil.bitcoin.BitCoinApi
 import com.omouravictor.ratesnow.data.network.hgbrasil.bitcoin.SourceRequestBitcoinModel
+import com.omouravictor.ratesnow.data.network.hgbrasil.rates.ApiService
 import com.omouravictor.ratesnow.utils.Resource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class BitCoinRepository @Inject constructor(
-    private val api: BitCoinApi, private val database: AppDataBase
+    private val api: ApiService, private val database: AppDataBase
 ) {
     suspend fun getAllFromApi(field: String): Resource<SourceRequestBitcoinModel> {
         return try {
-            val response = api.getBitCoin(field, BuildConfig.API_KEY_HG)
+            val response = api.getBitCoin(field)
             val result = response.body()
             if (response.isSuccessful && result != null)
                 Resource.Success(result)
